UPDATE / broke down BarHop into smaller components UPDATE / cleaned up some console logs
